+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - PR [#267](https://github.com/marinasundstrom/CheckedExceptions/pull/267) Add separate "Redundant catch clause" diagnostic to catch
 
+### Fixed
+
+- PR [#269](https://github.com/marinasundstrom/CheckedExceptions/pull/269) Silence declared exception from implicit object creation
+
 ## [2.0.3] - 2025-08-12
 
 ### Added